New feature and code enhancement (#99)
* Add support for multiple test paths and improve parsing

- Updated `MainCommand` to include a new required property `TestPaths` for multiple command line test paths.
- Modified `Run` method to iterate over `TestPaths` and print each path.
- Updated `launchSettings.json` to support new command line arguments for `TestPaths`.
- Refactored command line argument parsing in `ApplicationBuilder.cs` to use `ParseToType` instead of `Parse`.
- Introduced `ArrayTypeParser` for handling array types in command line parsing.
- Modified `ICommandLineTypeParser` interface to reflect new method signatures for improved type handling.

* Refactor ApplicationBuilder to use object representation

Replaced string representation of current values with object representation in the ApplicationBuilder class. This change enhances type safety and ensures consistency when setting default values and descriptions.

* Update default value handling in ApplicationBuilder

Refactor the logic for setting default values in `ApplicationBuilder.cs` to prioritize environment variables. The new implementation checks for a specified environment variable and uses its value if available. If the environment variable is not set, it falls back to comparing `defaultValue` with `currentValue` to determine the appropriate default. This replaces the previous logic that redundantly checked for the environment variable.

* Enumerable chnages

* Enumerable chnages

* Refactor Option instantiation in ApplicationBuilder

Updated the `Option` class instantiation to use `Option<string[]>` for enumerable property types. Removed debug output statements for cleaner code while preserving functionality.